+++ src/main/java/org/jax/mgi/snpindexer/indexes/SearchSNPIndexer.java
@@ -241,7 +241,7 @@ public class SearchSNPIndexer extends Indexer {
 
 			doc.setConsensussnp_accid(set.getString("consensussnp_accid"));
 			doc.setChromosome(set.getString("chromosome"));
-			doc.setStartcoordinate(set.getDouble("startcoordinate"));
+			doc.setStartcoordinate(set.getLong("startcoordinate"));
 			doc.setVarclass(variationMap.get(set.getString("_varclass_key")));
 
 			if (functionClassesMap.containsKey(set.getString("_object_key"))) {
